+++ Back/Dior.Library/Interfaces/UserInterface/Services/IAuditLogService.cs
@@ -7,10 +7,10 @@ namespace Dior.Library.Interfaces.UserInterface.Services
     public interface IAuditLogService
     {
         Task<IEnumerable<AuditLogDto>> GetAllAsync();
-        Task<AuditLogDto?> GetByIdAsync(int id);
+        Task<AuditLogDto?> GetByIdAsync(long id);
         Task<AuditLogDto> CreateAsync(CreateAuditLogDto createDto);
-        Task<List<AuditLogDto>> GetByEntityAsync(string entityType, int entityId);
-        Task<List<AuditLogDto>> GetByUserAsync(int userId);
+        Task<List<AuditLogDto>> GetByEntityAsync(string entityType, long entityId);
+        Task<List<AuditLogDto>> GetByUserAsync(long userId);
         Task<List<AuditLogDto>> GetRecentAsync(int count = 50);
     }
 }
