+++ Back/Dior.Library/Interfaces/UserInterface/Services/IRoleDefinitionPrivilegeService.cs
@@ -5,9 +5,9 @@ namespace Dior.Library.Interfaces.UserInterface.Services
     public interface IRoleDefinitionPrivilegeService
     {
         List<Dior.Library.Entities.RoleDefinitionPrivilege> GetList();
-        Dior.Library.Entities.RoleDefinitionPrivilege Get(int id);
+        Dior.Library.Entities.RoleDefinitionPrivilege Get(long id);
         long Add(Dior.Library.Entities.RoleDefinitionPrivilege entity, string editBy);
         void Set(Dior.Library.Entities.RoleDefinitionPrivilege entity, string editBy);
-        void Del(int id);
+        void Del(long id);
     }
 }
