+++ Back/Dior.Library/Interfaces/UserInterface/Services/IRoleService.cs
@@ -7,12 +7,12 @@ namespace Dior.Library.Interfaces.UserInterface.Services
     public interface IRoleService
     {
         Task<IEnumerable<RoleDefinitionDto>> GetAllAsync();
-        Task<RoleDefinitionDto?> GetByIdAsync(int id);
+        Task<RoleDefinitionDto?> GetByIdAsync(long id);
         Task<RoleDefinitionDto> CreateAsync(CreateRoleDefinitionDto createDto);
-        Task<bool> UpdateAsync(int id, UpdateRoleDefinitionDto updateDto);
-        Task<bool> DeleteAsync(int id);
-        Task<bool> ExistsAsync(int id);
+        Task<bool> UpdateAsync(long id, UpdateRoleDefinitionDto updateDto);
+        Task<bool> DeleteAsync(long id);
+        Task<bool> ExistsAsync(long id);
         Task<List<string>> GetRoleNamesAsync();
-        Task<List<PrivilegeDto>> GetRolePrivilegesAsync(int roleId);
+        Task<List<PrivilegeDto>> GetRolePrivilegesAsync(long roleId);
     }
 }
