+++ Back/Dior.Service/DAO/IDA_Privilege.cs
@@ -4,11 +4,11 @@ namespace Dior.Library.Service.DAO
 {
     public interface IDA_Privilege
     {
-        Privilege Get(int id);
+        Privilege Get(long id);
         List<Privilege> GetList();
         long Add(Privilege privilege, string editBy);
         void Set(Privilege privilege, string editBy);
-        void Del(int id);
+        void Del(long id);
         long Add(Entities.Privilege privilege, string editBy);
         void Set(Entities.Privilege privilege, string editBy);
     }
