+++ Back/Dior.Service/DAO/ITaskDao.cs
@@ -5,15 +5,15 @@ namespace Dior.Library.DAO
     public interface ITaskDao
     {
         List<TaskBO> GetAll();
-        TaskBO? GetById(int id);
+        TaskBO? GetById(long id);
         void Create(TaskBO task);
         void Update(TaskBO task);
-        void Delete(int id);
+        void Delete(long id);
 
-        List<TaskBO> GetTasksAssignedToUser(int userId);
-        List<TaskBO> GetTasksCreatedByUser(int userId);
+        List<TaskBO> GetTasksAssignedToUser(long userId);
+        List<TaskBO> GetTasksCreatedByUser(long userId);
         List<TaskBO> GetTasksByStatus(string status);
-        int GetTaskCountByStatusForUser(int userId, string status);
-        void UpdateTaskStatus(int taskId, string newStatus, string lastEditBy);
+        int GetTaskCountByStatusForUser(long userId, string status);
+        void UpdateTaskStatus(long taskId, string newStatus, string lastEditBy);
     }
 }
