+++ Back/Dior.Service/DAO/ITeamDao.cs
@@ -5,9 +5,9 @@ namespace Dior.Library.DAO
     public interface ITeamDao
     {
         List<Team> GetAll();
-        Team? GetById(int id);
+        Team? GetById(long id);
         void Create(Team team);
         void Update(Team team);
-        void Delete(int id);
+        void Delete(long id);
     }
 }
