+++ Back/Dior.Service/Services/Interfaces/IAuditLogService.cs
@@ -4,11 +4,14 @@ using System.Threading.Tasks;
 
 namespace Dior.Database.Services.Interfaces
 {
+    /// <summary>
+    /// Service de logs d'audit
+    /// </summary>
     public interface IAuditLogService
     {
         Task<IEnumerable<AuditLogDto>> GetAllAsync();
-        Task<AuditLogDto> GetByIdAsync(int id);
-        Task<IEnumerable<AuditLogDto>> GetByUserIdAsync(int userId);
+        Task<AuditLogDto> GetByIdAsync(long id);
+        Task<IEnumerable<AuditLogDto>> GetByUserIdAsync(long userId);
         Task<IEnumerable<AuditLogDto>> GetByTableNameAsync(string tableName);
         Task<AuditLogDto> CreateAsync(CreateAuditLogDto createAuditLogDto);
     }
