+++ Back/Dior.Service/Services/Interfaces/IRoleDefinitionService.cs
@@ -4,13 +4,16 @@ using System.Threading.Tasks;
 
 namespace Dior.Data.Services.Interfaces
 {
+    /// <summary>
+    /// Service de gestion des définitions de rôles
+    /// </summary>
     public interface IRoleDefinitionService
     {
         Task<IEnumerable<RoleDefinitionDto>> GetAllAsync();
-        Task<RoleDefinitionDto> GetByIdAsync(int id);
+        Task<RoleDefinitionDto> GetByIdAsync(long id);
         Task<RoleDefinitionDto> CreateAsync(CreateRoleDefinitionDto createRoleDefinitionDto);
-        Task<bool> UpdateAsync(int id, UpdateRoleDefinitionDto updateRoleDefinitionDto);
-        Task<bool> DeleteAsync(int id);
-        Task<IEnumerable<RoleDefinitionDto>> GetChildRolesAsync(int parentId);
+        Task<bool> UpdateAsync(long id, UpdateRoleDefinitionDto updateRoleDefinitionDto);
+        Task<bool> DeleteAsync(long id);
+        Task<IEnumerable<RoleDefinitionDto>> GetChildRolesAsync(long parentId);
     }
 }
