+++ Back/Dior.Service/Services/Interfaces/IUserAccessService.cs
@@ -4,13 +4,16 @@ using System.Threading.Tasks;
 
 namespace Dior.Library.Services.Interfaces
 {
+    /// <summary>
+    /// Service de gestion des UserAccess
+    /// </summary>
     public interface IUserAccessService
     {
         Task<IEnumerable<UserAccessDto>> GetAllAsync();
-        Task<UserAccessDto> GetByIdAsync(int id);
-        Task<IEnumerable<UserAccessDto>> GetByUserIdAsync(int userId);
+        Task<UserAccessDto> GetByIdAsync(long id);
+        Task<IEnumerable<UserAccessDto>> GetByUserIdAsync(long userId);
         Task<UserAccessDto> CreateAsync(CreateUserAccessDto createUserAccessDto);
-        Task<bool> UpdateAsync(int id, UpdateUserAccessDto updateUserAccessDto);
-        Task<bool> DeleteAsync(int id);
+        Task<bool> UpdateAsync(long id, UpdateUserAccessDto updateUserAccessDto);
+        Task<bool> DeleteAsync(long id);
     }
 }
